@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "build": "npm run lint && npm run clean && npm run webpack",
-    "webpack": "webpack --progress --colors --config webpack.config.js --output-path ./dist -p",
+    "webpack": "webpack --progress --config webpack.config.js --output-path ./dist ",
     "postbuild": "npm run package",
     "package": "tfx extension create --rev-version --manifest-globs ./vss-extension.json",
     "gallery-publish": "tfx extension publish --rev-version",
@@ -18,28 +18,29 @@
   "devDependencies": {
     "@types/react": "^16.9.2",
     "@types/react-dom": "^16.9.0",
-    "base64-inline-loader": "^1.1.1",
-    "css-loader": "^3.2.0",
-    "node-sass": "^4.12.0",
-    "promise-polyfill": "^8.1.3",
+    "base64-inline-loader": "^2.0.1",
+    "copy-webpack-plugin": "^11.0.0",
+    "css-loader": "^6.8.1",
+    "promise-polyfill": "^8.3.0",
     "requirejs": "^2.3.6",
-    "rimraf": "^3.0.0",
-    "sass-loader": "^8.0.0",
-    "style-loader": "^1.0.0",
-    "ts-loader": "^6.0.4",
-    "tslint": "^5.19.0",
-    "typescript": "^3.5.3",
-    "vss-web-extension-sdk": "^5.141.0",
-    "webpack": "^4.39.2",
-    "webpack-cli": "^3.3.7",
-    "webpack-dev-server": "^3.8.0"
+    "rimraf": "^5.0.1",
+    "sass": "^1.63.6",
+    "sass-loader": "^13.3.2",
+    "style-loader": "^3.3.3",
+    "ts-loader": "^9.4.4",
+    "tslint": "^5.20.1",
+    "typescript": "^5.1.6",
+    "url-loader": "^4.1.1",
+    "webpack": "^5.88.1",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^4.15.1"
   },
   "dependencies": {
-    "azure-devops-extension-api": "^1.154.1",
-    "azure-devops-extension-sdk": "^2.0.11",
-    "azure-devops-ui": "^1.157.0",
-    "moment": "^2.24.0",
-    "office-ui-fabric-react": "^7.33.0",
+    "@fluentui/react": "^8.110.10",
+    "azure-devops-extension-api": "^2.222.0",
+    "azure-devops-extension-sdk": "^3.1.2",
+    "azure-devops-ui": "^2.167.75",
+    "moment": "^2.29.4",
     "react": "^16.9.0",
     "react-dom": "^16.9.0"
   }
