@@ -1,8 +1,9 @@
 import "es6-promise/auto";
-import * as SDK from "azure-devops-extension-sdk";
-import { renderApp } from "./controls/app";
-import { applyFabricUITheme } from "./styles";
 import './main.scss';
+import { renderApp } from "../src/controls/app";
+import { applyFabricUITheme } from "./styles";
+import * as SDK from "azure-devops-extension-sdk";
+
 
 SDK.init();
 SDK.ready().then(async () => {
