@@ -1,37 +1,8 @@
-import { GitRestClient, PullRequestStatus, GitPullRequest, GitPullRequestSearchCriteria } from "azure-devops-extension-api/Git";
-import { IdentityRef } from "azure-devops-extension-api/WebApi";
-import { WorkItemTrackingRestClient, WorkItemExpand, WorkItemErrorPolicy, WorkItem } from "azure-devops-extension-api/WorkItemTracking";
+import { GitRestClient, PullRequestStatus, GitPullRequestSearchCriteria, GitCommitRef, GitPullRequest } from "azure-devops-extension-api/Git";
+import { WorkItemTrackingRestClient, WorkItem, WorkItemExpand, WorkItemErrorPolicy } from "azure-devops-extension-api/WorkItemTracking";
 import * as API from "azure-devops-extension-api";
+import { PullRequestRef, Issue } from "../releaseNotes";
 
-export class ZeroPullRequestRef {
-    repositoryId: string;
-}
-
-export class PullRequestRef {
-    id: number;
-    repositoryId: string;
-    title: string;
-    status: string;
-    creationDate: Date;
-    createdBy: IdentityRef;
-
-    constructor(fields?: Partial<PullRequestRef>) {
-        Object.assign(this, fields);
-    }
-}
-
-export class Issue {
-    href: string;
-    id: number;
-    type: string;
-    title: string;
-    tags: string[];
-    status: string;
-
-    constructor(fields?: Partial<Issue>) {
-        Object.assign(this, fields);
-    }
-}
 
 const TOP_PULLREQUESTS = 10;
 
@@ -64,11 +35,12 @@ const mapToReleaseNotesIssue = (wit: WorkItem): Issue =>
         status: (<string | undefined>wit.fields["System.State"] || "")
     });
 
+
 export class ReleaseNotesService {
     async getTopPullRequests(repositoryId: string): Promise<PullRequestRef[]> {
         const gitClient = API.getClient(GitRestClient);
         const searchCriteria = {
-            targetRefName: "refs/heads/master",
+            targetRefName: "refs/heads/main",
             status: PullRequestStatus.All,
             repositoryId: repositoryId,
             creatorId: "",
@@ -77,7 +49,7 @@ export class ReleaseNotesService {
             sourceRefName: "",
             sourceRepositoryId: "",
         } as GitPullRequestSearchCriteria;
-        
+
         const pullRequests = await gitClient.getPullRequests(
             repositoryId,
             searchCriteria,
@@ -99,12 +71,29 @@ export class ReleaseNotesService {
 
         const workItems = await this.getWorkItems(witClient, workItemsIds);
 
+        return this.getReleaseNotesIssues(workItems, witClient);
+    }
+
+    async getReleaseNotesIssuesBasedOnTags(commits: GitCommitRef[]): Promise<Issue[]> {
+        const witClient = API.getClient(WorkItemTrackingRestClient);
+
+        const commitsWorkItemsId = commits.flatMap(commit => commit.workItems.map(workItem => +workItem.id));
+
+        const workItems = await this.getWorkItems(witClient, commitsWorkItemsId);
+
+        return this.getReleaseNotesIssues(workItems, witClient);
+    }
+
+    private async getReleaseNotesIssues(workItems: WorkItem[], witClient: WorkItemTrackingRestClient): Promise<Issue[]> {
+
         const isTask = (wit: WorkItem): boolean => wit.fields["System.WorkItemType"] === "Task";
+
         const nonTaskWorkItems = workItems.filter(x => !isTask(x));
         const taskWorkItems = workItems.filter(x => isTask(x));
+
         const taskParentWorkItems = await this.getParentWorkItems(witClient, taskWorkItems, nonTaskWorkItems.map(x => x.id));
 
-        return [...nonTaskWorkItems, ...taskParentWorkItems].map(mapToReleaseNotesIssue).sort((a, b) => a.id - b.id);
+        return [...nonTaskWorkItems, ...taskParentWorkItems].map(mapToReleaseNotesIssue).sort((a, b) => b.id - a.id);
     }
 
     private async getParentWorkItems(witClient: WorkItemTrackingRestClient, taskWorkItems: WorkItem[], excludeWorkItems: number[]): Promise<WorkItem[]> {
@@ -129,15 +118,15 @@ export class ReleaseNotesService {
     }
 
     private async getPullRequestWorkItems(gitClient: GitRestClient, repositoryId: string, pullRequestId: number): Promise<number[]> {
-        var workItemRefs = await gitClient.getPullRequestWorkItemRefs(repositoryId, pullRequestId);
+        const workItemRefs = await gitClient.getPullRequestWorkItemRefs(repositoryId, pullRequestId);
         return workItemRefs.map(w => Number(w.url.split('/').pop()));
     }
 
     private async getCommentsWorkItems(gitClient: GitRestClient, repositoryId: string, pullRequestId: number): Promise<number[]> {
-        var commitRefs = await gitClient.getPullRequestCommits(repositoryId, pullRequestId);
+        const commitRefs = await gitClient.getPullRequestCommits(repositoryId, pullRequestId);
 
         return commitRefs.map(c => {
-            var match = /\b#[0-9]+\b/.exec(c.comment);
+            const match = /\b#[0-9]+\b/.exec(c.comment);
             return match ? match.map(x => Number(x)) : [];
         }).reduce((c, a) => c.concat(a), []);
     }
